JSON Export Format - Revision 2
Use nullable ints to make the exported blob slightly smaller.
Wrap the FromString() function in a try/catch block, if the input is invalid we give up and return null.
Store the power full name instead of separate Category, Set, Name values. This makes the following step easier.
Use a dictionary file to translate incorrect internal power names in the Mids database to the correct ones.
Write the correct level and relative level for enhancements.